Solve kata https://www.codewars.com/kata/55b051fac50a3292a9000025/train/java

+++ java/7_kyu/FilterTheNumber/Kata.java
@@ -8,6 +8,6 @@ public class Kata {
                 output.append(letter - 48);
             }
         }
-        return Integer.parseInt(output.toString());
+        return Long.parseLong(output.toString());
     }
 }
